+++ server/modules/media/controllers/index.js
@@ -9,7 +9,7 @@ import { authenticate } from '../../../api/middlewares';
 const router = Router();
 
 router.get('/', authenticate(), getSignedURL);
-router.get('/media-url', authenticate(), getMediaUrl);
+router.get('/url', authenticate(), getMediaUrl);
 router.get('/:id', authenticate(), getMediaById);
 
 export default router;
